+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Bump version of `schema` dependency to fix errors with newere Clickhouse versions now using `system.tables` table instead of `information_schema.tables` view. 
 
 ## v4.0.0-rc.3