+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+* Fixed the timestamp parsing in Clickhouse dialect.
+* Fixed the schema in the tutorial for clickhouse.
 * Bump version of `schema` dependency to fix errors with new Clickhouse versions now using `system.tables` table instead of `information_schema.tables` view.
 * Add `--network` flag to override the default value in the manifest or spkg
 
-## v4.0.0-rc.3
+## v4.0.0
 
 * Fix a critical bug breaking the reorg management when more than one row needs to be reverted.